requirements.txt
# Discord Bot
discord.py==2.3.2

# Spotify Integration
spotipy==2.23.0

# Environment Variables
python-dotenv==1.0.0

# Database
# sqlite3 is built into Python

# Logging
# logging is built into Python

# Web Server for Setup
Flask==3.0.0
gunicorn==21.2.0
